# Use new instance of stream in each benchmark run